# isp-framework/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DotMac ISP Framework",
    "description": "Comprehensive ISP management system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/portal-management/portal-management/accounts/admin": {
      "post": {
        "tags": [
          "portal-management",
          "Portal Management"
        ],
        "summary": "Create Portal Account Admin",
        "description": "Create a new Portal Account with admin privileges.",
        "operationId": "create_portal_account_admin_api_v1_portal_management_portal_management_accounts_admin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalAccountAdminCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portal-management/portal-management/accounts/{account_id}/admin": {
      "put": {
        "tags": [
          "portal-management",
          "Portal Management"
        ],
        "summary": "Update Portal Account Admin",
        "description": "Update Portal Account with admin privileges.",
        "operationId": "update_portal_account_admin_api_v1_portal_management_portal_management_accounts__account_id__admin_put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalAccountAdminUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortalAccountAdminCreate": {
        "properties": {
          "account_type": {
            "$ref": "#/components/schemas/PortalAccountType",
            "default": "customer"
          },
          "email_notifications": {
            "type": "boolean",
            "title": "Email Notifications",
            "default": true
          },
          "sms_notifications": {
            "type": "boolean",
            "title": "Sms Notifications",
            "default": false
          },
          "theme_preference": {
            "type": "string",
            "pattern": "^(light|dark|auto)$",
            "title": "Theme Preference",
            "default": "light"
          },
          "language_preference": {
            "type": "string",
            "maxLength": 10,
            "title": "Language Preference",
            "default": "en"
          },
          "timezone_preference": {
            "type": "string",
            "maxLength": 50,
            "title": "Timezone Preference",
            "default": "UTC"
          },
          "session_timeout_minutes": {
            "type": "integer",
            "maximum": 480.0,
            "minimum": 5.0,
            "title": "Session Timeout Minutes",
            "default": 30
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "portal_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Portal Id"
          },
          "status": {
            "$ref": "#/components/schemas/PortalAccountStatus",
            "default": "active"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password",
            "default": true
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "default": false
          },
          "phone_verified": {
            "type": "boolean",
            "title": "Phone Verified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "PortalAccountAdminCreate",
        "description": "Schema for admin creation of Portal Account."
      },
      "PortalAccountAdminUpdate": {
        "properties": {
          "email_notifications": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Notifications"
          },
          "sms_notifications": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sms Notifications"
          },
          "theme_preference": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(light|dark|auto)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Preference"
          },
          "language_preference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Preference"
          },
          "timezone_preference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone Preference"
          },
          "session_timeout_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Timeout Minutes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PortalAccountStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "must_change_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Must Change Password"
          },
          "email_verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified"
          },
          "phone_verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Verified"
          },
          "two_factor_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Two Factor Enabled"
          },
          "failed_login_attempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Login Attempts"
          },
          "security_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Notes"
          }
        },
        "type": "object",
        "title": "PortalAccountAdminUpdate",
        "description": "Schema for admin updates to Portal Account."
      },
...
      "PortalAccountResponse": {
        "properties": {
          "account_type": {
            "$ref": "#/components/schemas/PortalAccountType",
            "default": "customer"
          },
          "email_notifications": {
            "type": "boolean",
            "title": "Email Notifications",
            "default": true
          },
          "sms_notifications": {
            "type": "boolean",
            "title": "Sms Notifications",
            "default": false
          },
          "theme_preference": {
            "type": "string",
            "pattern": "^(light|dark|auto)$",
            "title": "Theme Preference",
            "default": "light"
          },
          "language_preference": {
            "type": "string",
            "maxLength": 10,
            "title": "Language Preference",
            "default": "en"
          },
          "timezone_preference": {
            "type": "string",
            "maxLength": 50,
            "title": "Timezone Preference",
            "default": "UTC"
          },
          "session_timeout_minutes": {
            "type": "integer",
            "maximum": 480.0,
            "minimum": 5.0,
            "title": "Session Timeout Minutes",
            "default": 30
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "portal_id": {
            "type": "string",
            "title": "Portal Id"
          },
          "status": {
            "$ref": "#/components/schemas/PortalAccountStatus"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "two_factor_enabled": {
            "type": "boolean",
            "title": "Two Factor Enabled"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          },
          "phone_verified": {
            "type": "boolean",
            "title": "Phone Verified"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "last_successful_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Login"
          },
          "password_changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Changed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "portal_id",
          "status",
          "tenant_id",
          "two_factor_enabled",
          "email_verified",
          "phone_verified",
          "must_change_password",
          "last_successful_login",
          "password_changed_at",
          "created_at",
          "updated_at"
        ],
        "title": "PortalAccountResponse",
        "description": "Schema for Portal Account response."
      },
      "PortalAccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "locked",
          "pending_activation",
          "deactivated"
        ],
        "title": "PortalAccountStatus",
        "description": "Portal account status enumeration."
      },
      "PortalAccountType": {
        "type": "string",
        "enum": [
          "customer",
          "technician",
          "reseller"
        ],
        "title": "PortalAccountType",
        "description": "Portal account type enumeration."
      },
...
    }
  }
}